Inclussão do meio de pagamento Transferencia, experimento de criação de pdf e refatoração de código

+++ app/Http/Controllers/Colaborador/MovimentoSalarioControllerExcluir.php
@@ -11,8 +11,14 @@ class MovimentoSalarioControllerExcluir extends Controller
     public function excluir($id) {
         try {
             $movimento = MovimentoSalario::find($id);
-            $movimento->delete();
-            return redirect()->back();
+            $competenciaAtual = Session()->get('sistema')->competencia();
+            if ($movimento->competencia->id == $competenciaAtual->id) {
+                $movimento->delete();
+                return redirect()->back();
+            } else {
+                throw new \Exception('Só é possivel excluir movimentos da competência atual');
+            }
+
         } catch (\Exception $exception) {
             return redirect()
                 ->back()
